refactor: separate slot
