+++ packages/blocky-core/src/index.ts
@@ -2,13 +2,18 @@ export * from "./view/editor";
 export * as markup from "./model/markup";
 export * from "./view/controller";
 export * from "./block/basic";
-export * as serialize from "./model/serialize";
 export { makeDefaultIdGenerator, type IdGenerator } from "./helper/idHelper";
 export { type BannerFactory, type BannerInstance } from "./view/bannerDelegate";
 export { type ToolbarFactory } from "./view/toolbarDelegate";
-export { CollaborativeCursor, type CollaborativeCursorOptions } from "./view/collaborativeCursors";
+export {
+  CollaborativeCursor,
+  type CollaborativeCursorOptions,
+} from "./view/collaborativeCursors";
 export { type IPlugin } from "./registry/pluginRegistry";
-export { setTextTypeForTextBlock, getTextTypeForTextBlock } from "./block/textBlock";
+export {
+  setTextTypeForTextBlock,
+  getTextTypeForTextBlock,
+} from "./block/textBlock";
 export {
   type CursorState,
   State as DocumentState,
@@ -24,4 +29,4 @@ export {
   TextType,
   DocNodeName,
 } from "./model";
-export { TextBlockName } from "./block/textBlock"
+export { TextBlockName } from "./block/textBlock";
