+++ packages/blocky-core/src/model/index.ts
@@ -1,4 +1,4 @@
-export { default as State, DocNodeName } from "./state";
+export { State, DocNodeName } from "./state";
 export { type CursorState } from "./cursor";
 export { TextType } from "./text";
 export * from "./tree";
